backend/requirements.txt
# --- Core Framework ---
fastapi==0.109.0
uvicorn==0.27.0
pydantic

# --- Image Processing & AI ---
face-recognition==1.3.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.4

# --- Utilities ---
python-multipart==0.0.6

# --- Admin UI ---
streamlit==1.31.0
pandas==2.2.0